rag/requirements.txt
evaluate
rouge_score
openai
psycopg2-binary
neo4j
qdrant-client
sentence-transformers
torch